Communities data is cleaned up on delete

+++ apps/web/graphql/pages/logic.ts
@@ -14,6 +14,10 @@ import { Media, User, Constants } from "@courselit/common-models";
 import { Domain } from "../../models/Domain";
 import { homePageTemplate } from "./page-templates";
 import { publishTheme } from "../themes/logic";
+import getDeletedMediaIds, {
+    extractMediaIDs,
+} from "@/lib/get-deleted-media-ids";
+import { deleteMedia } from "@/services/medialit";
 const { product, site, blogPage, communityPage, permissions, defaultPages } =
     constants;
 const { pageNames } = Constants;
@@ -120,6 +124,13 @@ export const updatePage = async ({
         return null;
     }
 
+    const deletedMediaIds = getDeletedMediaIds(
+        JSON.stringify(page.draftLayout || ""),
+        inputLayout || "",
+    );
+    const publishedLayoutMediaIds = extractMediaIDs(
+        JSON.stringify(page.layout ?? []),
+    );
     if (inputLayout) {
         try {
             let layout;
@@ -137,19 +148,6 @@ export const updatePage = async ({
             } catch (err) {
                 throw new Error(`${responses.invalid_layout}: ${err.message}`);
             }
-            // for (let widget of layout) {
-            //     if (widget.shared && widget.widgetId) {
-            //         ctx.subdomain.draftSharedWidgets[widget.name] =
-            //             Object.assign(
-            //                 {},
-            //                 ctx.subdomain.draftSharedWidgets[widget.name],
-            //                 widget,
-            //             );
-            //         widget.settings = undefined;
-            //     }
-            // }
-            // (ctx.subdomain as any).markModified("draftSharedWidgets");
-            // await (ctx.subdomain as any).save();
             const layoutWithSharedWidgetsSettings =
                 await copySharedWidgetsToDomain(layout, ctx.subdomain);
             page.draftLayout = layoutWithSharedWidgetsSettings;
@@ -168,6 +166,14 @@ export const updatePage = async ({
     }
     if (typeof robotsAllowed === "boolean") {
         page.draftRobotsAllowed = robotsAllowed;
+    }
+
+    const deletableMediaIds = Array.from(deletedMediaIds).filter(
+        (mediaId) => !publishedLayoutMediaIds.has(mediaId),
+    );
+
+    for (const mediaId of deletableMediaIds) {
+        await deleteMedia(mediaId);
     }
 
     try {
@@ -414,6 +420,20 @@ export const deletePage = async (
 
     if (defaultPages.includes(id)) {
         throw new Error(responses.action_not_allowed);
+    }
+
+    const page = (await PageModel.findOne({
+        domain: ctx.subdomain._id,
+        pageId: id,
+    }).lean()) as unknown as Page;
+
+    if (!page) {
+        throw new Error(responses.item_not_found);
+    }
+
+    const mediaToBeDeleted = extractMediaIDs(JSON.stringify(page));
+    for (const mediaId of Array.from(mediaToBeDeleted)) {
+        await deleteMedia(mediaId);
     }
 
     await PageModel.deleteOne({
